refactor/move alignment transfer code to alignment module
