+++ tests/alignment/commentary_transfer/data/P1/IA6E66F92/metadata.json
@@ -1,0 +1,34 @@
+{
+  "id": "IA6E66F92",
+  "title": {
+    "BO": "དགོངས་པ་རབ་གསལ་ལས་སེམས་བསྐྱེད་དྲུག་པ། ཤོ་ལོ་ཀ ༡-༦༤",
+    "EN": "Illuminating the Intent Chapter 6, verses 1 to 64"
+  },
+  "author": {
+    "BO": "ཙོང་ཁ་པ་བློ་བཟང་གྲགས་པ།",
+    "EN": "Tsongkhapa Lobzang Drakpa"
+  },
+  "imported": "2025-02-23T08:58:35.106650",
+  "source": "http://purl.bdrc.io/resource/WA1NLM688",
+  "toolkit_version": "0.0.1",
+  "parser": "DocxNumberListTranslationParser",
+  "initial_creation_type": "google_docx",
+  "language": "en",
+  "source_metadata": {},
+  "bases": [
+    {
+      "B8B3": {
+        "source_metadata": {
+          "total_segments": 9
+        },
+        "base_file": "B8B3.txt"
+      }
+    }
+  ],
+  "copyright": {
+    "status": "Unknown",
+    "notice": "",
+    "info_url": null
+  },
+  "licence": "Unknown"
+}
