+++ tests/alignment/commentary_transfer/data/P2/IC7760088/metadata.json
@@ -1,0 +1,36 @@
+{
+  "id": "IC7760088",
+  "title": {
+    "BO": "དགོངས་པ་རབ་གསལ་ལས་སེམས་བསྐྱེད་དྲུག་པ། ཤོ་ལོ་ཀ ༡-༦༤",
+    "EN": "Illuminating the Intent Chapter 6, verses 1 to 64"
+  },
+  "author": {
+    "BO": "ཙོང་ཁ་པ་བློ་བཟང་གྲགས་པ།",
+    "EN": "Tsongkhapa Lobzang Drakpa"
+  },
+  "imported": "2025-02-23T08:59:18.985325",
+  "source": "http://purl.bdrc.io/resource/WA1NLM688",
+  "toolkit_version": "0.0.1",
+  "parser": "DocxNumberListTranslationParser",
+  "initial_creation_type": "google_docx",
+  "language": "en",
+  "source_metadata": {
+    "version_of": "IA6E66F92"
+  },
+  "bases": [
+    {
+      "A389": {
+        "source_metadata": {
+          "total_segments": 10
+        },
+        "base_file": "A389.txt"
+      }
+    }
+  ],
+  "copyright": {
+    "status": "Unknown",
+    "notice": "",
+    "info_url": null
+  },
+  "licence": "Unknown"
+}
