+++ tests/alignment/commentary_transfer/data/P3/I77BD6EA9/base/BEC3.txt
@@ -1,0 +1,9 @@
+གཟུང་བ་མེད་པར་ཞེས་བྱ་བ་ལ་སོགས་པ་གསུངས་ཏེ། 
+འཆད་པར་འགྱུར་བའི་རིགས་པས་ཞེས་བྱ་བ་ནི་ཅི་ལྟར་རླུང་གིས་ཞེས་བྱ་བ་ལ་སོགས་པའོ། 
+ཐ་སྙད་ཙམ་དུ་ཡོད་པ་དོན་དམ་པར་ཡོད་པ་མ་ཡིན་པ་དེ་ལ་དངོས་པོ་བརྟགས་པར་ཡོད་པ་ཞེས་བྱ་ལ། 
+ཕྱི་རོལ་གྱི་དོན་མེད་ན་ཞེས་བྱ་བ་ནི་ཤེས་པ་ལས་ཐ་དད་དུ་གྱུར་པའི་སྔོན་པོ་ལ་སོགས་པའི་དོན་མེད་ནའོ།
+དེ་བསམ་བྱ་ཞེས་བྱ་བ་ནི་རྨི་ལམ་ན་དོན་མེད་པའི་ཤེས་པ་ཡོད་པ་ཡིན་ནོ་ཞེས་བྱ་བ་དེ་ཡང་བསམ་པར་བྱའོ། །ཡང་དེ་ཅི་ཞིག་ཅེ་ན་ཞེས་བྱ་བ་ནི། བསམ་པར་བྱ་བ་གང་ཞིག་ཡིན། །ཞེས་བྱ་བའོ། །
+གང་ཚེ་ཞེས་བྱ་བ་ལ་སོགས་པ་ནི་དབུ་མ་གང་གི་ཚེ་རྨི་ལམ་ན་ཡང་སེམས་མེད་པ་དེའི་ཚེ་རྨི་ལམ་གྱི་དཔེ་དེ་མ་གྲུབ་པ་ཡིན་ནོ་ཞེས་བྱ་བའི་ཐ་ཚིག་གོ། །དེ་ཉིད་བསྟན་པར་བྱ་བའི་ཕྱིར། དེ་ཚེ་ཁྱོད་ཀྱི་དཔེ་ཡོད་མིན། །ཞེས་བྱ་བ་གསུངས་ཏེ། 
+གཉིད་སད་པའི་གནས་སྐབས་ན་དྲན་པའི་སྒོ་ནས་གལ་ཏེ་རྨི་ལམ་གྱི་གནས་སྐབས་ན་རྣམ་པར་ཤེས་པ་ཡོད་པར་ཁས་ལེན་ན། 
+ཅི་ལྟར་ཞེས་བྱ་བ་ལ་སོགས་པ་གསུངས་ཏེ། ཅི་ལྟར་ཉམས་སུ་མྱོང་བ་ལ་ནི་ངས་མཐོང་ངོ་སྙམ་པའི་དྲན་པ་ཡོད་པ་དེ་བཞིན་དུ་ཡུལ་ཡང་ཁས་བླང་བར་བྱ་དགོས་སོ། །ཡང་ན་རྣམ་པར་ཤེས་པ་ཡང་ཡོད་པ་མ་ཡིན་ནོ་ཞེས་བྱ་བ་ནི་གལ་ཏེ་ཡུལ་དྲན་དུ་ཟིན་ཀྱང་ཡུལ་ཁས་མི་ལེན་ན་རྣམ་པར་ཤེས་པ་དྲན་ཡང་རྣམ་པར་ཤེས་པ་མེད་པར་ཁས་བླང་བའི་ཕྱིར་ཁོ་བོའི་ཕྱོགས་གྲུབ་པ་ཡིན་ནོ། 
+གཉིད་ཀྱིས་ཞེས་བྱ་བ་ལ་སོགས་པ་གསུངས་ཏེ། གཉིད་ལོག་པས་མིག་དང༌། རྣ་བ་དང༌། སྣ་དང༌།
