+++ tests/alignment/commentary_transfer/data/P3/I77BD6EA9/metadata.json
@@ -1,0 +1,27 @@
+{
+  "id": "I77BD6EA9",
+  "title": {
+    "BO": "དགོངས་པ་རབ་གསལ་ལས་སེམས་བསྐྱེད་དྲུག་པ། ཤོ་ལོ་ཀ ༡-༦༤",
+    "EN": "Illuminating the Intent Chapter 6, verses 1 to 64"
+  },
+  "author": {
+    "BO": "ཙོང་ཁ་པ་བློ་བཟང་གྲགས་པ།",
+    "EN": "Tsongkhapa Lobzang Drakpa"
+  },
+  "imported": "2025-02-23T09:07:36.218466",
+  "source": "http://purl.bdrc.io/resource/WA1NLM688",
+  "toolkit_version": "0.0.1",
+  "parser": "DocxSimpleCommentaryParser",
+  "initial_creation_type": "google_docx",
+  "language": "en",
+  "source_metadata": {
+    "version_of": "IA6E66F92"
+  },
+  "bases": [],
+  "copyright": {
+    "status": "Unknown",
+    "notice": "",
+    "info_url": null
+  },
+  "licence": "Unknown"
+}
